Concurrent-and-Parallel-Programming-in-Python/requirements.txt
aiohttp
async-timeout
attrs
beautifulsoup4
certifi
chardet
greenlet
idna
importlib-metadata
lxml
multidict
psycopg2-binary
PyYAML
requests
soupsieve
SQLAlchemy
typing-extensions
urllib3
yarl
zipp
tqdm
